refactor: moved devtools into a dev menu
